+++ src/documentation/pages/foundations-wip/accessibility/links-and-forms/examples/_name-role-value.vue
@@ -35,15 +35,15 @@ import { Vue } from 'vue-facing-decorator';
     return {
       breadcrumbs: [
         {
-          url: 'https://www.w3.org/TR/WCAG20',
-          text: 'WCAG 2.0 Guideline',
+          url: 'https://www.w3.org/TR/WCAG22/',
+          text: 'WCAG 2.2 Guideline',
         },
         {
-          url: 'https://www.w3.org/TR/WCAG20/#ensure-compat',
+          url: 'https://www.w3.org/TR/WCAG22/#compatible',
           text: '4.1 Compatible',
         },
         {
-          url: 'https://www.w3.org/TR/WCAG20/#ensure-compat-rsv',
+          url: 'https://www.w3.org/TR/WCAG22/#name-role-value',
           text: '4.1.2 Name, Role, Value',
         },
       ],
@@ -53,7 +53,7 @@ import { Vue } from 'vue-facing-decorator';
           mainText: 'Robust',
         },
         {
-          tooltipText: 'Requirements must be satisfied to pass WCAG 2.0 Level A Success Criteria.',
+          tooltipText: 'Requirements must be satisfied to pass WCAG 2.2 Level A Success Criteria.',
           mainText: 'Level A',
         },
       ],